Add another member of the PCG family; factor out common code.
