+++ pcgrandom/test/test_pcg_xsl_rr_v0.py
@@ -1,5 +1,5 @@
 """
-Tests for the PCG_XSH_RR_V0 generator.
+Tests for the PCG_XSL_RR_V0 generator.
 """
 from __future__ import division
 
@@ -9,35 +9,35 @@ import itertools
 import math
 import unittest
 
-from pcgrandom.pcg_xsh_rr_v0 import PCG_XSH_RR_V0
+from pcgrandom.pcg_xsl_rr_v0 import PCG_XSL_RR_V0
 
 
 # Sequences used for detecting reproducibility regressions.
 seq0_seed12345_die_rolls = [
-    2, 5, 5, 4, 1, 6, 4, 4, 2, 5, 2, 6, 6, 2, 3, 1, 2, 6, 2, 6]
+    1, 2, 4, 4, 1, 4, 2, 2, 6, 1, 6, 2, 4, 6, 6, 5, 1, 1, 2, 6]
 seq0_seed12345_uniforms = [
-    float.fromhex('0x1.5086103ef2a40p-2'),
-    float.fromhex('0x1.90a33cef220a7p-1'),
-    float.fromhex('0x1.2be0bbe709ca8p-3'),
-    float.fromhex('0x1.21bded2b2a972p-1'),
-    float.fromhex('0x1.0a20e3b6ce7f8p-2'),
-    float.fromhex('0x1.5307c0cf5db08p-2'),
-    float.fromhex('0x1.cdd6de4c7c725p-1'),
-    float.fromhex('0x1.65d8359c989eap-2'),
-    float.fromhex('0x1.ea43f297b8b18p-3'),
-    float.fromhex('0x1.0be283eba30c0p-2'),
+    float.fromhex('0x1.36feb111851b8p-3'),
+    float.fromhex('0x1.d0eab4d2dacd0p-3'),
+    float.fromhex('0x1.3153584d8fc39p-1'),
+    float.fromhex('0x1.2839f1285a962p-1'),
+    float.fromhex('0x1.dd8596e786ab0p-5'),
+    float.fromhex('0x1.295b84d5041f7p-1'),
+    float.fromhex('0x1.86227587a6c34p-3'),
+    float.fromhex('0x1.5ac08820e8fc8p-3'),
+    float.fromhex('0x1.bbdbd5409790bp-1'),
+    float.fromhex('0x1.1b964efcf0d0cp-3'),
 ]
 seq0_seed12345_choices = [
-    'HT', 'D2', 'CK', 'DJ', 'S7', 'C8', 'DJ', 'DT',
-    'HA', 'D4', 'HT', 'CT', 'C7',
+    'S7', 'S3', 'D9', 'DT', 'SJ', 'DT', 'S5', 'S6',
+    'C8', 'S7', 'C9', 'HK', 'DQ'
 ]
 seq0_seed12345_sample = [
-    'HT', 'C3', 'C8', 'DA', 'CA', 'D6', 'C6', 'H9',
-    'SK', 'H8', 'HJ', 'H2', 'H6',
+    'C8', 'S5', 'D6', 'DK', 'CK', 'S9', 'C2', 'HJ',
+    'SK', 'S7', 'D2', 'C6', 'C5',
 ]
 seq0_seed12345_shuffle = [
-    5, 19, 1, 14, 10, 8, 11, 15, 18, 12,
-    3, 7, 4, 2, 17, 0, 13, 6, 9, 16,
+    19, 1, 9, 5, 18, 2, 15, 8, 6, 3,
+    4, 0, 14, 10, 12, 11, 13, 16, 17, 7,
 ]
 
 
@@ -74,10 +74,10 @@ def count_samples_generated(gen):
         del gen._next_word
 
 
-class Test_PCG_XSH_RR_V0(unittest.TestCase):
+class Test_PCG_XSL_RR_V0(unittest.TestCase):
     def test_creation_without_seed(self):
-        gen1 = PCG_XSH_RR_V0()
-        gen2 = PCG_XSH_RR_V0()
+        gen1 = PCG_XSL_RR_V0()
+        gen2 = PCG_XSL_RR_V0()
 
         sample1 = [gen1.random() for _ in range(10)]
         sample2 = [gen2.random() for _ in range(10)]
@@ -87,7 +87,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.assertNotEqual(sample1, sample2)
 
     def test_seed_resets_gauss_state(self):
-        gen = PCG_XSH_RR_V0()
+        gen = PCG_XSL_RR_V0()
 
         gen.seed(2143)
         x1 = gen.random()
@@ -101,7 +101,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.assertEqual(y1, y2)
 
     def test_reproducibility(self):
-        gen = PCG_XSH_RR_V0(seed=12345, sequence=0)
+        gen = PCG_XSL_RR_V0(seed=12345, sequence=0)
         die_rolls = [gen.randint(1, 6) for _ in range(20)]
         self.assertEqual(
             die_rolls,
@@ -111,11 +111,11 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         # Assumes IEEE 754 binary64 floats. For other formats, we'd only expect
         # to get approximate equality here. If this ever actually breaks on
         # anyone's machine, please report to me and I'll fix it.
-        gen = PCG_XSH_RR_V0(seed=12345, sequence=0)
+        gen = PCG_XSL_RR_V0(seed=12345, sequence=0)
         uniforms = [gen.random() for _ in range(10)]
         self.assertEqual(uniforms, seq0_seed12345_uniforms)
 
-        gen = PCG_XSH_RR_V0(seed=12345, sequence=0)
+        gen = PCG_XSL_RR_V0(seed=12345, sequence=0)
         suits = 'SHDC'
         values = 'AKQJT98765432'
         cards = [suit+value for suit in suits for value in values]
@@ -134,8 +134,8 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.assertEqual(population, seq0_seed12345_shuffle)
 
     def test_sequence_default(self):
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345)
+        gen1 = PCG_XSL_RR_V0(seed=12345, sequence=0)
+        gen2 = PCG_XSL_RR_V0(seed=12345)
         self.assertEqual(gen1.getstate(), gen2.getstate())
 
     def test_independent_sequences(self):
@@ -146,8 +146,8 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         # normally distributed with mean 0 and standard deviation 1 /
         # (12*sqrt(N)). So we expect |V| to be at most 1 / (4*sqrt(N)) with
         # over 99% probability.
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345, sequence=1)
+        gen1 = PCG_XSL_RR_V0(seed=12345, sequence=0)
+        gen2 = PCG_XSL_RR_V0(seed=12345, sequence=1)
         N = 10000
         xs = [gen1.random() for _ in range(N)]
         ys = [gen2.random() for _ in range(N)]
@@ -157,20 +157,20 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
 
     def test_no_shared_state(self):
         # Get samples first from gen1, then from gen2.
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345, sequence=1)
+        gen1 = PCG_XSL_RR_V0(seed=12345, sequence=0)
+        gen2 = PCG_XSL_RR_V0(seed=12345, sequence=1)
         sample1_1 = [gen1.random() for _ in range(10)]
         sample2_1 = [gen2.random() for _ in range(10)]
 
         # Now in the opposite order: from gen2, then from gen1.
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345, sequence=1)
+        gen1 = PCG_XSL_RR_V0(seed=12345, sequence=0)
+        gen2 = PCG_XSL_RR_V0(seed=12345, sequence=1)
         sample2_2 = [gen2.random() for _ in range(10)]
         sample1_2 = [gen1.random() for _ in range(10)]
 
         # Now interleaved.
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345, sequence=1)
+        gen1 = PCG_XSL_RR_V0(seed=12345, sequence=0)
+        gen2 = PCG_XSL_RR_V0(seed=12345, sequence=1)
         sample1_3 = []
         sample2_3 = []
         for _ in range(10):
@@ -184,7 +184,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.assertEqual(sample2_1, sample2_3)
 
     def test_save_and_restore_state(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=27)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=27)
 
         # Generate some values.
         [gen.random() for _ in range(10)]
@@ -199,16 +199,16 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.assertEqual(samples2, samples3)
 
     def test_restore_state_from_different_generator(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=27)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=27)
         state = gen.getstate()
 
-        bad_version = 'pcgrandom.PCG_XSH_RR_V1'
+        bad_version = 'pcgrandom.PCG_XSL_RR_V1'
         bad_state = (bad_version,) + state[1:]
         with self.assertRaises(ValueError):
             gen.setstate(bad_state)
 
     def test_state_preserves_gauss(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=27)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=27)
 
         # Test a state with gauss_next = None
         state = gen.getstate()
@@ -225,14 +225,14 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.assertEqual(samples1, samples2)
 
     def test_getrandbits(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         k = 5
         samples = [gen.getrandbits(k) for _ in range(10000)]
         self.check_uniform(range(2**k), samples)
 
     def test_getrandbits_large(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         k = 379
         samples = [gen.getrandbits(k) for _ in range(1000)]
@@ -244,17 +244,17 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.assertGreaterEqual(max(samples), 2**378)
 
     def test_getrandbits_zero_bits(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
         result = gen.getrandbits(0)
         self.assertEqual(result, 0)
 
     def test_getrandbits_negative_bits(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
         with self.assertRaises(ValueError):
             gen.getrandbits(-1)
 
     def test_randrange_uniform(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         # Indirect test of our _randbelow override.
         n = 13
@@ -263,7 +263,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
 
     def test_randrange_one(self):
         # Corner case.
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
         state_before = gen.getstate()
 
         samples = [gen.randrange(1) for _ in range(10)]
@@ -273,25 +273,25 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.assertEqual(state_before, state_after)
 
     def test_randrange_of_wordsize(self):
-        # Using randrange(2**32) should consume exactly one
+        # Using randrange(2**64) should consume exactly one
         # sample each time.
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         with count_samples_generated(gen) as wordgen:
             for index in range(10000):
                 self.assertEqual(wordgen.call_count, index)
-                gen.randrange(2**32)
+                gen.randrange(2**64)
                 self.assertEqual(wordgen.call_count, index+1)
 
     def test_jumpahead(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
         # Generate samples, each sample consuming exactly one word
         # from the core generator.
-        samples = [gen.randrange(2**32) for _ in range(1000)]
+        samples = [gen.randrange(2**64) for _ in range(1000)]
 
         # Rewind, check we can produce the exact same samples.
         gen.jumpahead(-1000)
-        same_again = [gen.randrange(2**32) for _ in range(1000)]
+        same_again = [gen.randrange(2**64) for _ in range(1000)]
         self.assertEqual(samples, same_again)
 
         # Corner case: jumpahead(0) should work.
@@ -302,18 +302,18 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
 
         # Now jump around randomly within the collection of samples. Use a
         # separate generator for the positions to jump to.
-        posgen = PCG_XSH_RR_V0(seed=19733, sequence=22)
+        posgen = PCG_XSL_RR_V0(seed=19733, sequence=22)
 
         current_pos = 1000
         for _ in range(100):
             next_pos = posgen.randrange(1000)
             gen.jumpahead(next_pos - current_pos)
-            sample = gen.randrange(2**32)
+            sample = gen.randrange(2**64)
             self.assertEqual(sample, samples[next_pos])
             current_pos = next_pos + 1
 
     def test_randrange_float_arguments(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         with self.assertRaises(TypeError):
             gen.randrange(5.0)
@@ -328,7 +328,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
 
     def test_randrange_start_only(self):
         # See http://bugs.python.org/issue9379 for related discussion.
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
         samples = [gen.randrange(23) for _ in range(1000)]
         # Chance of getting only 22 of the possible 23 outcomes is < 1.2e-18.
         self.assertEqual(set(samples), set(range(23)))
@@ -339,7 +339,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
             gen.randrange(-5)
 
     def test_randrange_start_and_stop(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
         samples = [gen.randrange(-10, 13) for _ in range(1000)]
         self.assertEqual(set(samples), set(range(-10, 13)))
 
@@ -349,7 +349,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
             gen.randrange(21, 20)
 
     def test_randrange_start_stop_and_step(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         test_ranges = [
             (16, 37, 5),
@@ -377,7 +377,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
             gen.randrange(47, 47, 1)
 
     def test_choice(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         with self.assertRaises(IndexError):
             gen.choice([])
@@ -390,28 +390,28 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.check_uniform(seq, choices)
 
     def test_sample(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         samples = [tuple(gen.sample(range(4), 3)) for _ in range(10000)]
         population = list(itertools.permutations(range(4), 3))
         self.check_uniform(population, samples)
 
     def test_sample_set(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         s = set('ABCDEFG')
         sample = gen.sample(s, 3)
         self.assertLess(set(sample), s)
 
     def test_sample_dict(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         d = dict(a=1, b=2, c=3)
         with self.assertRaises(TypeError):
             gen.sample(d, 2)
 
     def test_sample_corner_cases(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
         self.assertEqual(gen.sample([], 0), [])
         with self.assertRaises(ValueError):
             self.assertEqual(gen.sample([], 1))
@@ -419,7 +419,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
             self.assertEqual(gen.sample([], -1))
 
     def test_shuffle(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
         population = list(range(13))
 
         result = gen.shuffle(population)
@@ -441,7 +441,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
 
     def test_shuffle_corner_cases(self):
         # shuffling a length 0 or 1 sequence shouldn't be a problem
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
         population = []
         gen.shuffle(population)
         self.assertEqual(population, [])
@@ -451,7 +451,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.assertEqual(population, [13])
 
     def test_choices(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         population = list(range(5))
 
@@ -468,7 +468,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
 
     def test_choices_subnormal_weights(self):
         # Corner case where random.Random triggers an IndexError.
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         population = list(range(5))
         weights = [1e-323, 0.0, 2e-323, 2e-323, 1e-323]
@@ -481,7 +481,7 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
         self.check_goodness_of_fit(dict(zip(population, weights)), sample)
 
     def test_choices_error_conditions(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         # Can't specify both weights and cum_weights.
         with self.assertRaises(TypeError):
@@ -522,17 +522,17 @@ class Test_PCG_XSH_RR_V0(unittest.TestCase):
     def test_count_samples_generated(self):
         # This is really a test for our count_samples_generated helper
         # rather than for the PRNG.
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
+        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
 
         [gen.randrange(13) for _ in range(14)]
         with count_samples_generated(gen) as wordgen:
             [gen.randrange(27) for _ in range(21)]
             self.assertEqual(wordgen.call_count, 21)
             [gen.random() for _ in range(10)]
-            self.assertEqual(wordgen.call_count, 41)
+            self.assertEqual(wordgen.call_count, 31)
 
         [gen.randrange(27) for _ in range(13)]
-        self.assertEqual(wordgen.call_count, 41)
+        self.assertEqual(wordgen.call_count, 31)
 
     def check_uniform(self, population, sample):
         """
